chore(#3063): code tidy
