+++ src/images_list.h
@@ -18,9 +18,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 *******************************************************/
 
 #include <wx/imaglist.h>
-#include <wx/sharedptr.h>
 #include <wx/arrstr.h>
-#include <map>
 
 enum img {
     HOUSE_PNG,
@@ -186,126 +184,6 @@ enum png {
     MAX_PNG
 };
 
-// SVG filename in Zip, the PNG enum to which it relates, the background color for the generated bitmap 0 = transparent
-static const std::map<std::string, std::pair<int, std::uint32_t>> iconName2enum = {
-    { "NEW_DB.svg", { NEW_DB, 0 } },
-    { "OPEN.svg", { OPEN, 0 } },
-    { "NEW_ACC.svg", { NEW_ACC, 0 } },
-    { "HOME.svg", { HOME, 0 } },
-    { "CATEGORY.svg", { CATEGORY, 0 } },
-    { "PAYEE.svg", { PAYEE, 0 } },
-    { "CURR.svg", { CURR, 0 } },
-    { "FILTER.svg", { FILTER, 0 } }, 
-    { "GRM.svg", { GRM, 0 } },
-    { "OPTIONS.svg", { OPTIONS, 0 } },
-    { "NEW_TRX.svg", { NEW_TRX, 0 } },
-    { "NEW_NEWS.svg", { NEW_NEWS, 0 } },
-    { "NEWS.svg", { NEWS, 0 } },
-    { "CURRATES.svg", { CURRATES, 0 } },
-    { "FULLSCREEN.svg", { FULLSCREEN, 0 } },
-    { "PRINT.svg", { PRINT, 0 } },
-    { "ABOUT.svg", { ABOUT, 0 } },
-    { "HELP.svg", { HELP, 0 } },
-
-    // Navigation
-    { "NAV_HOME.svg", { NAV_HOME, 0xFFFFFFFF } },
-    { "ALLTRANSACTIONS.svg", { ALLTRANSACTIONS, 0xFFFFFFFF } },
-    { "NAV_FILTER.svg", { NAV_FILTER, 0xFFFFFFFF } },
-    { "NAV_GRM.svg", { NAV_GRM, 0xFFFFFFFF } },
-    { "NAV_HELP.svg", { NAV_HELP, 0xFFFFFFFF } },
-    { "FAVOURITE.svg", { FAVOURITE, 0xFFFFFFFF } },
-    { "SAVINGS_NORMAL.svg", { SAVINGS_NORMAL, 0xFFFFFFFF } },
-    { "SAVINGS_CLOSED.svg", { SAVINGS_CLOSED, 0xFFFFFFFF } },
-    { "CC_NORMAL.svg", { CC_NORMAL, 0xFFFFFFFF } },
-    { "CC_CLOSED.svg", { CC_CLOSED, 0xFFFFFFFF } },
-    { "CASH_NORMAL.svg", { CASH_NORMAL, 0xFFFFFFFF } },
-    { "CASH_CLOSED.svg", { CASH_CLOSED, 0xFFFFFFFF } },
-    { "LOAN_ACC_NORMAL.svg", { LOAN_ACC_NORMAL, 0xFFFFFFFF } },
-    { "LOAN_ACC_CLOSED.svg", { LOAN_ACC_CLOSED, 0xFFFFFFFF } },
-    { "TERM_NORMAL.svg", { TERM_NORMAL, 0xFFFFFFFF } },
-    { "TERM_CLOSED.svg", { TERM_CLOSED, 0xFFFFFFFF } },
-    { "STOCKS_NORMAL.svg", { STOCKS_NORMAL, 0xFFFFFFFF } },
-    { "STOCKS_CLOSED.svg", { STOCKS_CLOSED, 0xFFFFFFFF } },
-    { "ASSET_NORMAL.svg", { ASSET_NORMAL, 0xFFFFFFFF } },
-    { "ASSET_CLOSED.svg", { ASSET_CLOSED, 0xFFFFFFFF } },
-    { "RECURRING.svg", { RECURRING, 0xFFFFFFFF } },
-    { "BUDGET.svg", { BUDGET, 0xFFFFFFFF } },
-    { "PIE_CHART.svg", { PIE_CHART, 0xFFFFFFFF } },
-     
-    // Status
-    { "RECONCILED.svg", { RECONCILED, 0 } },
-    { "DUPLICATE_STAT.svg", { DUPLICATE_STAT, 0 } },
-    { "FOLLOW_UP.svg", { FOLLOW_UP, 0 } },
-    { "VOID_STAT.svg", { VOID_STAT, 0 } },
-    { "PROFIT.svg", { PROFIT, 0 } },
-    { "LOSS.svg", { LOSS, 0 } },
-    { "LED_OFF.svg", { LED_OFF, 0 } },
-    { "LED_GREEN.svg", { LED_GREEN, 0 } },
-    { "LED_YELLOW.svg", { LED_YELLOW, 0 } },
-    { "LED_RED.svg", { LED_RED, 0 } },
-    { "RUN.svg", { RUN, 0 } },
-    { "RUN_AUTO.svg", { RUN_AUTO, 0 } },
-
-    // Assets
-    { "PROPERTY.svg", { PROPERTY, 0 } },
-    { "CAR.svg", { CAR, 0 } },
-    { "HOUSEHOLD_OBJ.svg", { HOUSEHOLD_OBJ, 0 } },
-    { "ART.svg", { ART, 0 } },
-    { "JEWELLERY.svg", { JEWELLERY, 0 } },
-    { "CASH.svg", { CASH, 0 } },
-    { "OTHER.svg", { OTHER, 0 } },
-
-    //Controls
-    { "CLEAR.svg", { CLEAR, 0 } },
-    { "CLIP.svg", { CLIP, 0 } },
-    { "DOWNARROW.svg", { DOWNARROW, 0 } },
-    { "UPARROW.svg", { UPARROW, 0 } },
-    { "LEFTARROW.svg", { LEFTARROW, 0 } },
-    { "RIGHTARROW.svg", { RIGHTARROW, 0 } },
-    { "IMPORT.svg", { IMPORT, 0 } },
-    { "MORE_OPTIONS.svg", { MORE_OPTIONS, 0 } },
-    { "RELOCATION.svg", { RELOCATION, 0 } },
-    { "SAVE.svg", { SAVE, 0 } },
-    { "TRANSFILTER.svg", { TRANSFILTER, 0 } },
-    { "TRANSFILTER_ACTIVE.svg", { TRANSFILTER_ACTIVE, 0 } },
-    { "TRXNUM.svg", { TRXNUM, 0 } },
-    { "UPDATE.svg", { UPDATE, 0 } },
-    { "WEB.svg", { WEB, 0 } },
-
-    // Settings
-    { "GENERAL.svg", { GENERAL, 0 } },
-    { "VIEW.svg", { VIEW, 0 } },
-    { "ATTACHMENTS.svg", { ATTACHMENTS, 0 } },
-    { "NETWORK.svg", { NETWORK, 0 } },
-    { "OTHERS.svg", { OTHERS, 0 } },
-
-    // Extra Account Icons
-    { "ACC_MONEY.svg", { ACC_MONEY, 0xFFFFFFFF } },
-    { "ACC_EURO.svg", { ACC_EURO, 0xFFFFFFFF } },
-    { "ACC_FLAG.svg", { ACC_FLAG, 0xFFFFFFFF } },
-    { "ACC_ABOUT.svg", { ACC_ABOUT, 0xFFFFFFFF } },
-    { "ACC_COINS.svg", { ACC_COINS, 0xFFFFFFFF } },
-    { "ACC_CLOCK.svg", { ACC_CLOCK, 0xFFFFFFFF } },
-    { "ACC_CAT.svg", { ACC_CAT, 0xFFFFFFFF } },
-    { "ACC_DOG.svg", { ACC_DOG, 0xFFFFFFFF } },
-    { "ACC_TREES.svg", { ACC_TREES, 0xFFFFFFFF } },
-    { "ACC_HOURGLASS.svg", { ACC_HOURGLASS, 0xFFFFFFFF } },
-    { "ACC_WORK.svg", { ACC_WORK, 0xFFFFFFFF } },
-    { "ACC_PAYPAL.svg", { ACC_PAYPAL, 0xFFFFFFFF } },
-    { "ACC_WALLET.svg", { ACC_WALLET, 0xFFFFFFFF } },
-    { "ACC_RUBIK.svg", { ACC_RUBIK, 0xFFFFFFFF } },
-
-    // Other
-    { "EMPTY.svg", { EMPTY, 0 } }
-};
-
-static bool iconsLoaded;
-static wxSharedPtr<wxBitmap> programIcons16[MAX_PNG];
-static wxSharedPtr<wxBitmap> programIcons24[MAX_PNG];
-static wxSharedPtr<wxBitmap> programIcons32[MAX_PNG];
-static wxSharedPtr<wxBitmap> programIcons48[MAX_PNG];
-
-static wxSharedPtr<wxArrayString> themes;
 wxArrayString getThemes();
 
 wxImageList* navtree_images_list();
